refactor: move menu api to the API package
