+++ packages/api/src/menu-context.ts
@@ -1,0 +1,25 @@
+/**********************************************************************
+ * Copyright (C) 2025 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ***********************************************************************/
+
+export enum MenuContext {
+  DASHBOARD_IMAGE = 'dashboard/image',
+  DASHBOARD_CONTAINER = 'dashboard/container',
+  DASHBOARD_POD = 'dashboard/pod',
+  DASHBOARD_COMPOSE = 'dashboard/compose',
+  DASHBOARD_CONTAINER_CONNECTION = 'dashboard/container-connection',
+}
